feat(package): expose agentv sdk facade
