+++ apps/cli/src/config.ts
@@ -1,0 +1,4 @@
+export {
+  defineConfig,
+  type AgentVTsConfig as AgentVConfig,
+} from '@agentv/core';
