+++ apps/cli/src/provider.ts
@@ -1,0 +1,20 @@
+export {
+  createBuiltinProviderRegistry,
+  createProvider,
+  resolveAndCreateProvider,
+  resolveDelegatedProviderDefinition,
+  resolveProviderDefinition,
+  resolveProviderDefinitionEnvironments,
+  type EnvLookup,
+  type Message,
+  type Provider,
+  type ProviderDefinition,
+  type ProviderFactoryFn,
+  type ProviderKind,
+  type ProviderRequest,
+  type ProviderResponse,
+  type ProviderStreamCallbacks,
+  type ProviderTokenUsage,
+  type ResolvedProviderBackend,
+  type ToolCall,
+} from '@agentv/core';
