+++ apps/cli/src/sdk.ts
@@ -1,0 +1,9 @@
+/**
+ * User-facing SDK helpers for `agentv/sdk`.
+ *
+ * This leaf entrypoint intentionally avoids importing CLI command modules so
+ * script graders and TypeScript eval files can import helpers without loading
+ * the command tree.
+ */
+export * from '@agentv/sdk';
+export * from './config.js';
