+++ packages/core/src/script-grader.ts
@@ -1,0 +1,3 @@
+export * from './evaluation/script-grader-runtime.js';
+export * from './evaluation/script-grader-schemas.js';
+export * from './evaluation/vitest-workspace-grader.js';
