+++ packages/sdk/package.json
@@ -1,7 +1,8 @@
 {
   "name": "@agentv/sdk",
   "version": "5.3.1-next.1",
-  "description": "Evaluation SDK for AgentV - build custom assertions and script graders",
+  "private": true,
+  "description": "Internal implementation package for AgentV TypeScript SDK helpers",
   "type": "module",
   "repository": {
     "type": "git",
@@ -21,19 +22,16 @@
     }
   },
   "scripts": {
-    "prepublishOnly": "node -e \"if(process.env.ALLOW_PUBLISH!=='1'){console.error('ERROR: Use bun run publish:next, then bun run promote:latest');process.exit(1)}\"",
-    "build:deps": "bun --cwd ../.. --filter @agentv/core build",
-    "build": "bun run build:deps && tsup",
+    "build": "tsup",
     "dev": "tsup --watch",
-    "typecheck": "bun run build:deps && tsc --noEmit",
+    "typecheck": "tsc --noEmit",
     "lint": "biome check .",
     "format": "biome format --write .",
     "fix": "biome check --write .",
-    "test": "bun run build:deps && bun test"
+    "test": "bun test"
   },
-  "files": ["dist", "README.md"],
   "dependencies": {
-    "@agentv/core": "5.3.1-next.1",
+    "@agentv/core": "workspace:*",
     "yaml": "^2.8.3",
     "zod": "^3.23.8"
   }
